// src/shared/types.ts
export type SizeVariant = 'small' | 'medium' | 'large';

export type ButtonVariant =
  | 'filled'
  | 'outline'
  | 'filled-warning'
  | 'outline-warning'
  | 'subtle'
  | 'subtle-warning';

export type IconButtonVariant =
  | 'filled'
  | 'outline'
  | 'subtle'
  | 'warning'
  | 'subtle-warning'
  | 'filled-warning'
  | 'outline-warning';

export type FloatingButtonVariant = 'filled' | 'outline';

export type TagVariant = 'neutral' | 'positive' | 'negative' | 'accent1' | 'accent2';

export type ToastVariant = 'success' | 'info' | 'warning' | 'error';

export type BodyTextVariant =
  | 'body-large'
  | 'body'
  | 'body-small'
  | 'caption'
  | 'overline'
  | 'label';

export type HeadingLevel = 1 | 2 | 3 | 4 | 5 | 6;